requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
openai
pinecone
